requirements.txt
django==2.0.6
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
djangorestframework==3.9.0
Markdown==3.0.1
django-ckeditor==5.6.1
djangorestframework-jwt==1.11.0
coreapi==2.3.3
num2words=0.5.9